API

AttributeGraphs.addedge! — Method

addedge!(ag::AttributeGraph, args...) -> Bool

Add an edge in the graph and update attributes and keys/indices

AttributeGraphs.addedgeattr! — Method

addedgeattr!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:AbstractDict{Kr, Vr}},
    e::AbstractEdge,
    m::Integer,
    v
)

Add edge attribute v for edge e and multiplicity m.

AttributeGraphs.addedgeattr! — Method

addedgeattr!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:AbstractDict{Kr, Vr}},
    e::AbstractEdge,
    v
) -> Any

Add edge attribute v for edge e

AttributeGraphs.addedgeattr! — Method

addedgeattr!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:AbstractDict{Kr, Vr}},
    s::Integer,
    d::Integer,
    m::Integer,
    v
)

Add edge attribute v for edge defined by source s, destination d and multiplicity m.

AttributeGraphs.addedgeattr! — Method

addedgeattr!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:AbstractDict{Kr, Vr}},
    s::Integer,
    d::Integer,
    v
) -> Any

Add edge attribute v for edge defined by source s and destination d

AttributeGraphs.addgraphattr! — Method

addgraphattr!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E, R<:AbstractDict{Kr, Vr}},
    k,
    v
) -> Any

Add a graph attribute

AttributeGraphs.addvertex! — Method

addvertex!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V<:(AbstractVector), E, R}
) -> Bool

Add a vertex in the graph and push a missing value to the attributes

AttributeGraphs.addvertexattr! — Method

addvertexattr!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E},
    k::Integer,
    v
) -> Any

Add vertex attribute

AttributeGraphs.edge_attr — Method

edge_attr(
    mg::AttributeGraphs.AbstractAttributeGraph{T<:Integer, G<:AbstractGraph}
) -> Any

Get the data structure for all edge attributes

AttributeGraphs.getedgeattr — Method

getedgeattr(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:(AbstractDict{Kr})},
    s::Integer,
    d::Integer,
    m::Integer
) -> Any

Get edge attribute for edge defined by source s and destination d and multiplicity m

AttributeGraphs.getedgeattr — Method

getedgeattr(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:(AbstractDict{Kr})},
    s::Integer,
    d::Integer
) -> Any

Get edge attribute for edge defined by source s and destination d

AttributeGraphs.getedgeattr — Method

getedgeattr(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:AbstractDict},
    e::AbstractEdge,
    m::Integer
) -> Any

Get edge attribute for edge defined edge e with multiplicity m

AttributeGraphs.getedgeattr — Method

getedgeattr(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:AbstractDict},
    e::AbstractEdge
) -> Any

Get edge attribute for edge defined by e

AttributeGraphs.getgraphattr — Method

getgraphattr(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E, R<:(AbstractDict{Kr})},
    k
) -> Any

Get graph attribute

AttributeGraphs.getvertexattr — Method

getvertexattr(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E},
    k::Integer
) -> Any

Get vertex attribute

AttributeGraphs.getwgraph — Method

getwgraph(
    mg::AttributeGraphs.AbstractAttributeGraph{T<:Integer, G<:AbstractGraph}
) -> Any

Get the wrapped graph

AttributeGraphs.graph_attr — Method

graph_attr(
    mg::AttributeGraphs.AbstractAttributeGraph{T<:Integer, G<:AbstractGraph}
) -> Any

Get the data structure for all graph attributes

AttributeGraphs.hasedgeattr — Method

hasedgeattr(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:(AbstractDict{Kr})},
    ktup::Tuple
) -> Any

Check if edge has attribute for edge defined by the tuple ktup

AttributeGraphs.hasedgeattr — Method

hasedgeattr(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:(AbstractDict{Kr})},
    args::T<:Integer...
) -> Any

Check if edge has attribute for edge defined by the arguments args given as a tuple

AttributeGraphs.hasedgeattr — Method

hasedgeattr(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:AbstractDict},
    e::AbstractEdge,
    m::Integer
) -> Any

Check if edge has attribute for edge e and multiplicity m

AttributeGraphs.hasedgeattr — Method

hasedgeattr(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:AbstractDict},
    e::AbstractEdge
) -> Any

Check if edge has attribute for edge e

AttributeGraphs.hasgraphattr — Method

hasgraphattr(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E, R<:(AbstractDict{Kr})},
    k
) -> Any

Check if graph has this attribute

AttributeGraphs.hasvertexattr — Method

hasvertexattr(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E},
    k::Integer
) -> Bool

Check if vertex has this attribute

AttributeGraphs.remedge! — Method

remedge!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V<:(AbstractVector), E<:(AbstractDict{<:Tuple{T<:Integer, T<:Integer, T<:Integer}}), R},
    s::Integer,
    d::Integer
)
remedge!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V<:(AbstractVector), E<:(AbstractDict{<:Tuple{T<:Integer, T<:Integer, T<:Integer}}), R},
    s::Integer,
    d::Integer,
    m::Integer
)

Remove an edge from the graph and update attributes and keys/indices

AttributeGraphs.remedgeattr! — Method

remedgeattr!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:(AbstractDict{Kr})},
    e::AbstractEdge,
    m::Integer
) -> Any

Remove edge attribute for edge e and multiplicity m

AttributeGraphs.remedgeattr! — Method

remedgeattr!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:(AbstractDict{Kr})},
    e::AbstractEdge
) -> Any

Remove edge attribute for edge e

AttributeGraphs.remedgeattr! — Method

remedgeattr!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:(AbstractDict{Kr})},
    s::Integer,
    d::Integer,
    m::Integer
) -> Any

Remove edge attribute for edge defined by source s, destination d and multiplicity m.

AttributeGraphs.remedgeattr! — Method

remedgeattr!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E<:(AbstractDict{Kr})},
    s::Integer,
    d::Integer
) -> Any

Remove edge attribute for edge defined by source s and destination d

AttributeGraphs.remgraphattr! — Method

remgraphattr!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E, R<:(AbstractDict{Kr})},
    k
) -> Any

Remove a graph attribute

AttributeGraphs.remvertex! — Method

remvertex!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V<:(AbstractVector), E<:(AbstractDict{<:Tuple{T<:Integer, T<:Integer, T<:Integer}}), R},
    u::Integer
) -> Bool

Remove a vertex from the graph and update all attributes and keys/indices

AttributeGraphs.remvertexattr! — Method

remvertexattr!(
    ag::AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E},
    k::Integer
) -> Union{Missing, Nothing}

Remove vertex attribute

AttributeGraphs.vertex_attr — Method

vertex_attr(
    mg::AttributeGraphs.AbstractAttributeGraph{T<:Integer, G<:AbstractGraph}
) -> Any

Get the data structure for all vertex attributes

AttributeGraphs.AttributeGraph — Type

struct AttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}, V, E, R} <: AttributeGraphs.AbstractAttributeGraph{T<:Integer, G<:AbstractGraph{T<:Integer}}

graph::AbstractGraph{T} where T<:Integer: the wrapped graph

vertex_attr::Any: the attributes for the vertices

edge_attr::Any: the attributes for the edges

graph_attr::Any: the attributes for the graph

If the type for the vertices attributes and the edges attributes is a AbstractVector then there is a default behavior.

AttributeGraphs.AttributeGraph — Method

AttributeGraph(
    gr::AbstractGraph{T<:Integer};
    vertex_type,
    edge_type,
    graph_type
) -> AttributeGraph{T, _B, Missing, Missing, Missing} where {T<:Integer, _B<:AbstractGraph{T}}

Construct a AttributeGraph.

Keywords

vertex_attr_type: The element type for the vertex attributes

edge_attr_type: The element type for the edge attributes

graph_attr_type: The element type for the graph attributes

All types need to be able to call an empty constructor.

Index

AttributeGraphs.AttributeGraph

AttributeGraphs.AttributeGraph

AttributeGraphs.addedge!

AttributeGraphs.addedgeattr!

AttributeGraphs.addedgeattr!

AttributeGraphs.addedgeattr!

AttributeGraphs.addedgeattr!

AttributeGraphs.addgraphattr!

AttributeGraphs.addvertex!

AttributeGraphs.addvertexattr!

AttributeGraphs.edge_attr

AttributeGraphs.getedgeattr

AttributeGraphs.getedgeattr

AttributeGraphs.getedgeattr

AttributeGraphs.getedgeattr

AttributeGraphs.getgraphattr

AttributeGraphs.getvertexattr

AttributeGraphs.getwgraph

AttributeGraphs.graph_attr

AttributeGraphs.hasedgeattr

AttributeGraphs.hasedgeattr

AttributeGraphs.hasedgeattr

AttributeGraphs.hasedgeattr

AttributeGraphs.hasgraphattr

AttributeGraphs.hasvertexattr

AttributeGraphs.remedge!

AttributeGraphs.remedgeattr!

AttributeGraphs.remedgeattr!

AttributeGraphs.remedgeattr!

AttributeGraphs.remedgeattr!

AttributeGraphs.remgraphattr!

AttributeGraphs.remvertex!

AttributeGraphs.remvertexattr!

AttributeGraphs.vertex_attr